rename functions, break out colors for charts
