+++ ntd/prep_data_utils.py
@@ -96,24 +96,3 @@ def extra_annual_rtpa_splitting(row):
         return row.rtpa_name_split
     else:
         return row.rtpa_name
-
-
-# Define color palette here - easier to switch to DDS ones
-# if needed, without importing yet another dependency
-CALITP_CATEGORY_BRIGHT_COLORS = [
-    "#2EA8CE",  # darker blue
-    "#EB9F3C",  # orange
-    "#F4D837",  # yellow
-    "#51BF9D",  # green
-    "#8CBCCB",  # lighter blue
-    "#9487C0",  # purple
-]
-
-CALITP_CATEGORY_BOLD_COLORS = [
-    "#136C97",  # darker blue
-    "#E16B26",  # orange
-    "#F6BF16",  # yellow
-    "#00896B",  # green
-    "#7790A3",  # lighter blue
-    "#5B559C",  # purple
-]
