+++ ntd/update_vars.py
